[svn-r24596] Description:     Initial set of revisions, as preparing to bring feature back to trunk after review.
Tested on:
    Mac OSX/64 10.9.1 (amazon) w/C++, FORTRAN, thread-safe & build-all
    (h5committest not required for this branch)

+++ test/earray.c
@@ -2803,7 +2803,7 @@ main(void)
 
     /* Seed random #'s */
     curr_time = HDtime(NULL);
-    HDsrandom((unsigned long)curr_time);
+    HDsrandom((unsigned)curr_time);
 
     /* Create an empty file to retrieve size */
     {
